Fixed many more warnings that occur with gcc4;
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@289 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ test/VariantTest/VariantTest.cpp
@@ -117,8 +117,8 @@ void Test( const char* pszString, EPdfDataType eDataType, const char* pszExpecte
         RAISE_ERROR( ePdfError_TestFailed );
     }
 
-    printf("   -> Parsed Length    : %li (%i)\n", lLen, strlen(pszExpected) );
-    if( lLen != strlen( pszExpected ) )
+    printf("   -> Parsed Length    : %li (%i)\n", lLen, static_cast<int>(strlen(pszExpected)) );
+    if( lLen != static_cast<int>(strlen( pszExpected )) )
     {
         RAISE_ERROR( ePdfError_TestFailed );
     }
